requirements.txt
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.3.0
boto3==1.37.11
botocore==1.37.11
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.11
fastapi-users==14.0.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
makefun==1.15.6
motor==3.7.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pwdlib==0.2.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.2
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
